Tables and histograms
I kept the first table because it had the total number of variables at the top. Also all data is available for this data set which makes it very nice and tidy. The next table is the mean of each variable which is important. However, the next two tables are the max and min but those could also be the average or the median both would be easy fixes I just was not sure sorry. Next we have the histograms there are seven of them one for each variable. The frequencies are shown on the left side and then the shape of the histogram is what we will be annualizing for project. I do not know if we need each histogram but they are all coded for individually so very easy to delete. All titles are there but those can also be changed if you would like :). Also I kept everything that was there before at the bottom.